Re-adding GENie scripts.
CMake may be used by a lot of people, but it's an incredibly bloated build systems. Since this is a very lightweight library, I don't want to force people to bloat down their build setups by requiring CMake to be installed. GENie can be deployed with just an executable.

I'll probably still maintain CMakeLists.txt for ease of use by people using it.